# Speed up /api/v3/jobs: replace ROW_NUMBER() first-volume subquery with MIN(JobMediaId) join